Adding the Cities views. Code in need of documentation

+++ api/v1/views/states.py
@@ -34,8 +34,10 @@ def get_state(state_id):
 @app_views.route('/states/<state_id>', methods=['DELETE'], strict_slashes=False)
 def delete_state(state_id):
     state = storage.get(State, state_id)
+    if not state:
+        abort(404)
     storage.delete(state)
-    return jsonify({})
+    return make_response(jsonify({}), 200)
 
 @app_views.route('/states/<state_id>', methods=['PUT'], strict_slashes=False)
 def update_state(state_id):
